cannot read from unsigned if type is potentially signed

+++ fflas-ffpack/field/modular-balanced-double.h
@@ -227,7 +227,7 @@ namespace FFPACK
 
 		std::istream &read (std::istream &is, Element &x) const
 		{
-			unsigned long tmp;
+			double tmp;
 			is >> tmp;
 			init(x,tmp);
 			return is;
