+++ fflas-ffpack/field/modular-balanced-int32.h
@@ -185,7 +185,7 @@ namespace FFPACK
 
 		std::istream &read (std::istream &is, Element &x) const
 		{
-			double tmp;
+			long int tmp;
 			is >> tmp;
 			init(x,tmp);
 			return is;
